removed some views

+++ flask_backend/flaskModels.py
@@ -7,8 +7,8 @@ class Transactions(db.Model):
     stock_ticker= db.Column(db.String(1000))
     quantity = db.Column(db.Integer)
     action = db.Column(db.String(1000))
-    price = db.Column(db.Double)
-    total_price = db.Column(db.Double) #quantity times price
+    price = db.Column(db.Float)
+    total_price = db.Column(db.Float) #quantity times price
     date_time = db.Column(db.DateTime(timezone=True), default=func.now())
     user_id = db.Column(db.Integer, db.ForeignKey('user.id'))
